style: organize type definition files
