--- src/common/types/AddPeopleProps.ts
@@ -1,8 +1,0 @@
-type AddPeopleProps = {
-  id: number;
-  name: string;
-  number: string;
-  isEditormode: boolean;
-};
-
-export default AddPeopleProps;
