--- src/common/types/BtnType.ts
@@ -1,9 +1,0 @@
-type BtnType = {
-  url?: string;
-  text: string;
-  backgroundColor: string;
-  textColor: string;
-  isBorder: boolean;
-};
-
-export default BtnType;
