--- src/common/types/PostList.ts
@@ -1,50 +1,0 @@
-type PostList = {
-  imgUrl: string;
-  title: string;
-  content: string;
-};
-
-type PostCommentProp = {
-  username: string;
-  desc: string;
-  time: string;
-  site: string;
-};
-
-interface replyType {
-  bid: string;
-  reply_content: string;
-  parentUserName: string;
-  reply_creator: {
-    avatar: string;
-    studentid: string;
-    username: string;
-  };
-  reply_pos: string;
-  reply_time: string;
-}
-
-interface creatorType {
-  username: string;
-  avatar: string;
-  studentid: string;
-}
-
-interface responseType {
-  bid: string;
-  commented_pos: string;
-  commented_time: string;
-  content: string;
-  creator: creatorType;
-  isLike: string;
-  likeNum: number;
-  replyNum: number;
-  reply: replyType[];
-}
-
-interface PostCommentProps extends PostCommentProp {
-  res: PostCommentProp[];
-}
-
-export default PostList;
-export { PostCommentProps, responseType, replyType, creatorType };
