--- src/common/types/User.ts
@@ -1,18 +1,0 @@
-interface user {
-  id?: number;
-  avatar: string;
-  username: string;
-  school: string;
-  college?: string;
-  studentid: string;
-}
-
-/*
-头像：string
-用户名：string
-学校：string
-学院：string
-学号：string
-*/
-
-export default user;
