--- src/common/types/picturecut.ts
@@ -1,8 +1,0 @@
-type PictureCutProps = {
-    isvisible: boolean;
-    imgUrl: string[];
-    setImgUrl: (value: string[]) => void;
-    setIsVisible: (value: boolean) => void;
-    studentId: string;
-}
-export default PictureCutProps;
