+++ src/common/types/userTypes.ts
@@ -1,0 +1,15 @@
+// 用户信息存储类型
+export interface User {
+  /** 唯一标识 */
+  id?: number;
+  /** 头像url */
+  avatar: string;
+  /** 用户名 */
+  username: string;
+  /** 学校 */
+  school: string;
+  /** 学院 */
+  college?: string;
+  /** 学号 */
+  studentid: string;
+}
