+++ src/modules/CommentActionSheet/index.tsx
@@ -1,0 +1,2 @@
+import CommentActionSheet from './components/CommentActionSheet';
+export default CommentActionSheet;
