--- src/modules/CommentOperation/index.tsx
@@ -1,2 +1,0 @@
-import CommentOperation from './components/commentOperation';
-export default CommentOperation;
