+++ src/modules/picturecut/components/picturecut.tsx
@@ -2,7 +2,7 @@ import { View, Image, Canvas } from '@tarojs/components';
 import Taro from '@tarojs/taro';
 import { useState, useEffect, useRef } from 'react';
 import './style.scss';
-import PictureCutProps from '@/common/types/picturecut';
+import { PictureCutProps } from '@/common/types';
 import { fetchToQiniu } from '../../../common/api/qiniu';
 import post from '@/common/api/post';
 
@@ -47,7 +47,7 @@ const PictureCut: React.FC<PictureCutProps> = ({
     height: 200,
     scale: 1,
     minScale: 0.5,
-    maxScale: 3
+    maxScale: 3,
   });
 
   const [touchState, setTouchState] = useState<TouchState>({
@@ -58,13 +58,13 @@ const PictureCut: React.FC<PictureCutProps> = ({
     lastDistance: 0,
     isMoving: false,
     isScaling: false,
-    initialDistance: 0
+    initialDistance: 0,
   });
 
   const [imageInfo, setImageInfo] = useState<ImageInfo>({
     width: 0,
     height: 0,
-    aspectRatio: 1
+    aspectRatio: 1,
   });
 
   const [containerSize, setContainerSize] = useState({ width: 0, height: 0 });
@@ -85,7 +85,7 @@ const PictureCut: React.FC<PictureCutProps> = ({
           const containerWidth = res[0].width;
           const containerHeight = res[0].height;
           setContainerSize({ width: containerWidth, height: containerHeight });
-          
+
           Taro.getImageInfo({
             src: imgUrl[0],
             success: (info) => {
@@ -93,11 +93,11 @@ const PictureCut: React.FC<PictureCutProps> = ({
               setImageInfo({
                 width: info.width,
                 height: info.height,
-                aspectRatio
+                aspectRatio,
               });
 
               const cropSize = Math.min(containerWidth, containerHeight) * 0.8;
-              
+
               let initialScale = 1;
               if (aspectRatio >= 1) {
                 initialScale = cropSize / info.width;
@@ -108,7 +108,7 @@ const PictureCut: React.FC<PictureCutProps> = ({
               const displayHeight = info.height * initialScale;
               const cropBoxX = containerWidth / 2;
               const cropBoxY = containerHeight / 2;
-              
+
               const imageX = cropBoxX - displayWidth / 2;
               const imageY = cropBoxY - displayHeight / 2;
 
@@ -119,9 +119,9 @@ const PictureCut: React.FC<PictureCutProps> = ({
                 height: cropSize,
                 scale: initialScale,
                 minScale: initialScale * 0.5,
-                maxScale: initialScale * 3
+                maxScale: initialScale * 3,
               });
-            }
+            },
           });
         }
       });
@@ -147,7 +147,7 @@ const PictureCut: React.FC<PictureCutProps> = ({
 
   const handleTouchStart = (e: any) => {
     const touches = e.touches;
-    
+
     if (touches.length === 1) {
       setTouchState({
         startX: touches[0].clientX,
@@ -157,16 +157,16 @@ const PictureCut: React.FC<PictureCutProps> = ({
         lastDistance: 0,
         isMoving: true,
         isScaling: false,
-        initialDistance: 0
+        initialDistance: 0,
       });
     } else if (touches.length === 2) {
       const dx = touches[0].clientX - touches[1].clientX;
       const dy = touches[0].clientY - touches[1].clientY;
       const distance = Math.sqrt(dx * dx + dy * dy);
-      
+
       const centerX = (touches[0].clientX + touches[1].clientX) / 2;
       const centerY = (touches[0].clientY + touches[1].clientY) / 2;
-      
+
       setTouchState({
         startX: centerX,
         startY: centerY,
@@ -175,7 +175,7 @@ const PictureCut: React.FC<PictureCutProps> = ({
         lastDistance: distance,
         isMoving: false,
         isScaling: true,
-        initialDistance: distance
+        initialDistance: distance,
       });
     }
   };
@@ -187,57 +187,56 @@ const PictureCut: React.FC<PictureCutProps> = ({
 
     animationFrameRef.current = requestAnimationFrame(() => {
       const touches = e.touches;
-      
+
       if (touchState.isMoving && touches.length === 1) {
         const deltaX = touches[0].clientX - touchState.lastX;
         const deltaY = touches[0].clientY - touchState.lastY;
-        
-        setCropState(prev => ({
+
+        setCropState((prev) => ({
           ...prev,
           x: prev.x + deltaX,
-          y: prev.y + deltaY
+          y: prev.y + deltaY,
         }));
-        
-        setTouchState(prev => ({
+
+        setTouchState((prev) => ({
           ...prev,
           lastX: touches[0].clientX,
-          lastY: touches[0].clientY
+          lastY: touches[0].clientY,
         }));
-        
       } else if (touchState.isScaling && touches.length === 2) {
         const dx = touches[0].clientX - touches[1].clientX;
         const dy = touches[0].clientY - touches[1].clientY;
         const currentDistance = Math.sqrt(dx * dx + dy * dy);
-        
+
         const centerX = (touches[0].clientX + touches[1].clientX) / 2;
         const centerY = (touches[0].clientY + touches[1].clientY) / 2;
-        
+
         if (touchState.initialDistance > 0) {
           const scaleChange = currentDistance / touchState.initialDistance;
           const newScale = Math.max(
             cropState.minScale,
             Math.min(cropState.maxScale, cropState.scale * scaleChange)
           );
-          
+
           const scaleCenterX = (centerX - cropState.x) / cropState.scale;
           const scaleCenterY = (centerY - cropState.y) / cropState.scale;
-          
+
           const newX = centerX - scaleCenterX * newScale;
           const newY = centerY - scaleCenterY * newScale;
-          
-          setCropState(prev => ({
+
+          setCropState((prev) => ({
             ...prev,
             x: newX,
             y: newY,
-            scale: newScale
+            scale: newScale,
           }));
         }
-        
-        setTouchState(prev => ({
+
+        setTouchState((prev) => ({
           ...prev,
           lastX: centerX,
           lastY: centerY,
-          initialDistance: currentDistance
+          initialDistance: currentDistance,
         }));
       }
     });
@@ -252,7 +251,7 @@ const PictureCut: React.FC<PictureCutProps> = ({
       lastDistance: 0,
       isMoving: false,
       isScaling: false,
-      initialDistance: 0
+      initialDistance: 0,
     });
   };
 
@@ -297,12 +296,16 @@ const PictureCut: React.FC<PictureCutProps> = ({
         const canvas = res[0].node;
         const ctx = canvas.getContext('2d');
         const dpr = Taro.getSystemInfoSync().pixelRatio;
-        
-        console.log('Canvas配置:', { width: cropState.width * dpr, height: cropState.height * dpr, dpr });
+
+        console.log('Canvas配置:', {
+          width: cropState.width * dpr,
+          height: cropState.height * dpr,
+          dpr,
+        });
         canvas.width = cropState.width * dpr;
         canvas.height = cropState.height * dpr;
         ctx.scale(dpr, dpr);
-        
+
         ctx.clearRect(0, 0, cropState.width, cropState.height);
 
         if (cropImageRef.current) {
@@ -310,7 +313,7 @@ const PictureCut: React.FC<PictureCutProps> = ({
           cropImageRef.current.onerror = null;
           cropImageRef.current.src = '';
         }
-        
+
         const img = canvas.createImage();
         cropImageRef.current = img;
         img.src = imgUrl[0];
@@ -325,48 +328,54 @@ const PictureCut: React.FC<PictureCutProps> = ({
           }
           cleanupCropResources();
         }, 10000);
-        
+
         img.onload = () => {
           clearTimeout(imgLoadTimeout);
           console.log('图片加载成功:', { width: img.width, height: img.height });
           try {
             const cropBoxCenterX = containerSize.width / 2;
             const cropBoxCenterY = containerSize.height / 2;
-            
+
             const cropBoxLeft = cropBoxCenterX - cropState.width / 2;
             const cropBoxTop = cropBoxCenterY - cropState.height / 2;
-            
+
             const offsetX = cropBoxLeft - cropState.x;
             const offsetY = cropBoxTop - cropState.y;
-            
+
             const sourceX = Math.max(0, offsetX / cropState.scale);
             const sourceY = Math.max(0, offsetY / cropState.scale);
-            
-            const sourceWidth = Math.min(
-              cropState.width / cropState.scale,
-              img.width - sourceX
-            );
-            const sourceHeight = Math.min(
-              cropState.height / cropState.scale,
-              img.height - sourceY
-            );
+
+            const sourceWidth = Math.min(cropState.width / cropState.scale, img.width - sourceX);
+            const sourceHeight = Math.min(cropState.height / cropState.scale, img.height - sourceY);
 
             console.log('绘制图片参数:', {
-              sourceX, sourceY, sourceWidth, sourceHeight,
-              destX: 0, destY: 0, destWidth: cropState.width, destHeight: cropState.height
+              sourceX,
+              sourceY,
+              sourceWidth,
+              sourceHeight,
+              destX: 0,
+              destY: 0,
+              destWidth: cropState.width,
+              destHeight: cropState.height,
             });
-            
+
             ctx.drawImage(
               img,
-              sourceX, sourceY, sourceWidth, sourceHeight,
-              0, 0, cropState.width, cropState.height
+              sourceX,
+              sourceY,
+              sourceWidth,
+              sourceHeight,
+              0,
+              0,
+              cropState.width,
+              cropState.height
             );
 
             const canvasTimeout = setTimeout(() => {
               console.error('Canvas导出超时');
               cleanupCropResources();
             }, 10000);
-            
+
             Taro.canvasToTempFilePath({
               canvas,
               fileType: 'png',
@@ -376,29 +385,29 @@ const PictureCut: React.FC<PictureCutProps> = ({
                 console.log('裁剪成功，临时文件路径:', tempRes.tempFilePath);
                 try {
                   console.log('裁剪成功，临时文件路径:', tempRes.tempFilePath);
-                  
+
                   const url = await fetchToQiniu(tempRes.tempFilePath);
                   console.log('七牛云返回URL:', url);
-                  
+
                   const response = await post('/user/avatar', {
                     avatar_url: url,
                     studentid: studentId,
                   });
-                  
+
                   if (response.msg === 'success') {
                     setImgUrl([tempRes.tempFilePath]);
                     setIsVisible(false);
                   } else {
-                    Taro.showToast({ 
-                      title: '上传失败: ' + (response.msg || '未知错误'), 
-                      icon: 'none' 
+                    Taro.showToast({
+                      title: '上传失败: ' + (response.msg || '未知错误'),
+                      icon: 'none',
                     });
                   }
                 } catch (uploadError) {
                   console.error('上传失败:', uploadError);
-                  Taro.showToast({ 
-                    title: '上传失败', 
-                    icon: 'none' 
+                  Taro.showToast({
+                    title: '上传失败',
+                    icon: 'none',
                   });
                 } finally {
                   setIsLoading(false);
@@ -408,7 +417,7 @@ const PictureCut: React.FC<PictureCutProps> = ({
                 console.error('导出失败:', error);
                 Taro.showToast({ title: '导出失败', icon: 'none' });
                 setIsLoading(false);
-              }
+              },
             });
           } catch (error) {
             console.error('裁剪过程出错:', error);
@@ -416,7 +425,7 @@ const PictureCut: React.FC<PictureCutProps> = ({
             setIsLoading(false);
           }
         };
-        
+
         img.onerror = () => {
           console.error('图片加载失败');
           Taro.showToast({ title: '图片加载失败', icon: 'none' });
@@ -439,7 +448,7 @@ const PictureCut: React.FC<PictureCutProps> = ({
       </View>
 
       <View className="picturecut-content">
-        <View 
+        <View
           className="picturecut-crop-area"
           style={{
             width: `${cropState.width}px`,
@@ -461,7 +470,7 @@ const PictureCut: React.FC<PictureCutProps> = ({
           </View>
         </View>
 
-        <View 
+        <View
           className="picturecut-image-container"
           onTouchStart={handleTouchStart}
           onTouchMove={handleTouchMove}
@@ -472,27 +481,27 @@ const PictureCut: React.FC<PictureCutProps> = ({
             transformOrigin: '0 0',
           }}
         >
-          <Image 
-            src={imgUrl[0]} 
+          <Image
+            src={imgUrl[0]}
             className="picturecut-image"
             mode="widthFix"
-            style={{ 
+            style={{
               width: `${imageInfo.width}px`,
               height: `${imageInfo.height}px`,
             }}
           />
         </View>
-        
-        <Canvas 
+
+        <Canvas
           id="cropCanvas"
           className="picturecut-canvas"
-          style={{ 
-            width: `${cropState.width}px`, 
+          style={{
+            width: `${cropState.width}px`,
             height: `${cropState.height}px`,
             position: 'absolute',
             left: '-9999px',
             top: '-9999px',
-            opacity: 0 
+            opacity: 0,
           }}
           type="2d"
         />
@@ -502,8 +511,8 @@ const PictureCut: React.FC<PictureCutProps> = ({
           <View className="picturecut-btn picturecut-btn-cancel" onClick={handleCancel}>
             取消
           </View>
-          <View 
-            className="picturecut-btn picturecut-btn-confirm" 
+          <View
+            className="picturecut-btn picturecut-btn-confirm"
             onClick={handleCrop}
             style={{ opacity: isLoading ? 0.6 : 1 }}
           >
@@ -513,7 +522,7 @@ const PictureCut: React.FC<PictureCutProps> = ({
       </View>
 
       {isLoading && (
-        <View 
+        <View
           style={{
             position: 'absolute',
             top: 0,
