+++ src/subpackage/postAdd/index.tsx
@@ -11,7 +11,7 @@ import usePostStore from '@/store/PostStore';
 import useUserStore from '@/store/userStore';
 import post from '@/common/api/post';
 import get from '@/common/api/get';
-import LabelForm from '@/common/types/LabelForm';
+import { LabelForm } from '@/common/types';
 import Taro from '@tarojs/taro';
 
 const Index = () => {
@@ -54,24 +54,24 @@ const Index = () => {
         title: '请上传图片',
         icon: 'none',
       });
-    }else if (title.length === 0) {
+    } else if (title.length === 0) {
       Taro.showToast({
         title: '请输入标题',
         icon: 'none',
       });
-    }else{
-    setIsShowDraft(false);
-    const postInfo = { introduce, showImg: pageImgUrl, studentid, title };
-    console.log(postInfo);
-    
-    post('/post/create', postInfo)
-      .then((res) => {
-        console.log(res);
-        switchTab({ url: '/pages/postHome/index' });
-      })
-      .catch((err) => {
-        console.log(err);
-      });     
+    } else {
+      setIsShowDraft(false);
+      const postInfo = { introduce, showImg: pageImgUrl, studentid, title };
+      console.log(postInfo);
+
+      post('/post/create', postInfo)
+        .then((res) => {
+          console.log(res);
+          switchTab({ url: '/pages/postHome/index' });
+        })
+        .catch((err) => {
+          console.log(err);
+        });
     }
   };
 
@@ -87,64 +87,64 @@ const Index = () => {
     backgroundColor: '#CF79FA',
     textColor: '#FFFEFF',
     isBorder: false,
-  }
+  };
 
   return (
     <>
-    {load&&
-      <View className="addblog-introduce">
-        <View className="addblog-introduce-container">
-          <View className="addblog-introduce-container-title">{count}/1000</View>
-          <View className="addblog-introduce-container-content">
-            <Input
-              style={'font-size: 44rpx;color: #170A1E;font-family: SimHei;'}
-              className="addblog-introduce-container-content-title"
-              value={title}
-              onInput={(e) => setTitle(e.detail.value)}
-              placeholderClass="addblog-introduce-container-content-title-placeholder"
-              placeholder="清晰名称能更好地让人注意哦~"
-            ></Input>
-            <Textarea
-              className="addblog-introduce-container-content-desc"
-              value={introduce}
-              onInput={(e) => {
-                setIntroduce(e.detail.value);
-                setCount(e.detail.value.length);
-              }}
-              placeholderClass="addblog-introduce-container-content-desc-placeholder"
-              placeholder="为了让大家更好地了解该活动，请介绍一下活动亮点， 活动流程和注意事项等内容......"
-            ></Textarea>
-            <View className="addblog-introduce-container-content-pic">
-              {pageImgUrl &&
-                pageImgUrl.map((item, index) => (
-                  <Picture
-                    key={index}
-                    src={item}
-                    isShowDelete={true}
-                    imgUrl={pageImgUrl}
-                    setImgUrl={setPageImgUrl}
-                  />
-                ))}
-              <View
-                className="addblog-introduce-container-content-pic-addblog"
-                onClick={() => setIsShowAlbum(true)}
-              >
-                +
+      {load && (
+        <View className="addblog-introduce">
+          <View className="addblog-introduce-container">
+            <View className="addblog-introduce-container-title">{count}/1000</View>
+            <View className="addblog-introduce-container-content">
+              <Input
+                style={'font-size: 44rpx;color: #170A1E;font-family: SimHei;'}
+                className="addblog-introduce-container-content-title"
+                value={title}
+                onInput={(e) => setTitle(e.detail.value)}
+                placeholderClass="addblog-introduce-container-content-title-placeholder"
+                placeholder="清晰名称能更好地让人注意哦~"
+              ></Input>
+              <Textarea
+                className="addblog-introduce-container-content-desc"
+                value={introduce}
+                onInput={(e) => {
+                  setIntroduce(e.detail.value);
+                  setCount(e.detail.value.length);
+                }}
+                placeholderClass="addblog-introduce-container-content-desc-placeholder"
+                placeholder="为了让大家更好地了解该活动，请介绍一下活动亮点， 活动流程和注意事项等内容......"
+              ></Textarea>
+              <View className="addblog-introduce-container-content-pic">
+                {pageImgUrl &&
+                  pageImgUrl.map((item, index) => (
+                    <Picture
+                      key={index}
+                      src={item}
+                      isShowDelete={true}
+                      imgUrl={pageImgUrl}
+                      setImgUrl={setPageImgUrl}
+                    />
+                  ))}
+                <View
+                  className="addblog-introduce-container-content-pic-addblog"
+                  onClick={() => setIsShowAlbum(true)}
+                >
+                  +
+                </View>
               </View>
             </View>
           </View>
-        </View>
-        <View className="addblog-introduce-floor">
-          <View className="addblog-introduce-floor-draft" onClick={() => setIsShowDraft(true)}>
-            <Image src={draft} mode="widthFix" style={{ width: '60rpx' }}></Image>
-            <View className="addblog-introduce-floor-draft-text">存草稿</View>
+          <View className="addblog-introduce-floor">
+            <View className="addblog-introduce-floor-draft" onClick={() => setIsShowDraft(true)}>
+              <Image src={draft} mode="widthFix" style={{ width: '60rpx' }}></Image>
+              <View className="addblog-introduce-floor-draft-text">存草稿</View>
+            </View>
+            <View className="addblog-introduce-floor-btn" onClick={() => handleConfirm()}>
+              {pageImgUrl.length > 0 ? <Button {...btn} /> : <Button {...btnDisabled} />}
+            </View>
           </View>
-          <View className="addblog-introduce-floor-btn" onClick={() => handleConfirm()}>
-            {pageImgUrl.length > 0 ? <Button {...btn} /> : <Button {...btnDisabled} />}
-          </View>
         </View>
-      </View>    
-    }
+      )}
 
       {isShowDraft && (
         <DraftWinodw
